dataSyncFunction/requirements.txt
azure-functions
pandas
requests
sqlalchemy
pyodbc
xlrd
azure-storage==0.36.0



